Refactored classes and training loops
training loops are now fully contained inside the Session class defined in rlnn-refactored.py. Agents are built on top of the abstract Player class defined in tictactoegame.py, and so any arbitrary subclass can be used for the training runs. Old code is still there incase it ever needs to be referenced again.